# apiGateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_serj213_bookServiceApi_internal_domain.Book:
    properties:
      author:
        type: string
      categoryId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  internal_http.BookResponse:
    properties:
      author:
        type: string
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  internal_http.GetBooksResponseOk:
    properties:
      books:
        items:
          $ref: '#/definitions/internal_http.BookResponse'
        type: array
      status:
        type: string
    type: object
  internal_http.ResponseOkBody:
    properties:
      data:
        $ref: '#/definitions/github_com_serj213_bookServiceApi_internal_domain.Book'
      status:
        type: string
    type: object
host: localhost:8083
info:
  contact: {}
  description: This is a sample server for a book store.
  title: Book Store API
  version: "1.0"
paths:
  /book/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/internal_http.ResponseOkBody'
      summary: Получить книгу
      tags:
      - books
  /book/delete/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/internal_http.ResponseOkBody'
      summary: Удалить книгу
      tags:
      - books
  /book/update:
    put:
      consumes:
      - application/json
      description: обновить данные о книге
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/internal_http.BookResponse'
      summary: Обновить книгу
      tags:
      - books
  /books:
    get:
      consumes:
      - application/json
      description: Получить все книги из системы
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/internal_http.GetBooksResponseOk'
      summary: Получить список книг
      tags:
      - books
swagger: "2.0"